setup script work

+++ setup.sh
@@ -3,57 +3,21 @@
 echo "Beginning Tauchan installation prerequisites setup."
 echo "Installation order: GraphicsMagick, Redis, PostgreSQL"
 sleep 3
-_cwd = "$PWD"
-sudo service postgresql stop
-sudo sh -c 'echo "deb http://apt.postgresql.org/pub/repos/apt/ `lsb_release -cs`-pgdg main" >> /etc/apt/sources.list.d/pgdg.list'
-wget --quiet -O - https://www.postgresql.org/media/keys/ACCC4CF8.asc | sudo apt-key add -
-sudo apt-get update
-
-echo "------------------------------------------"
-echo "Installing the GraphicsMagick dependencies..."
-echo "Dependency downloading can take 5-10 minutes to complete. Please be patient."
-echo "------------------------------------------"
-sleep 3
-mkdir -p /tmp/gm
-wget ftp://ftp.graphicsmagick.org/pub/GraphicsMagick/1.3/GraphicsMagick-1.3.25.tar.gz -O /tmp/gm-1.3.25.tar.gz
-tar -xzf /tmp/gm-latest.tar.gz -C /tmp/gm --strip-components=1 && unlink /tmp/gm-1.3.25.tar.gz
-cd /tmp/gm && ./configure
-make
-make check
-sudo make install
-
-echo "----------------------------------"
-echo "Installing the Redis dependencies..."
-echo "Dependency tests can take 5-10 minutes to complete."
-echo "There will be a couple times where user interaction is needed. Please stand by."
-echo "----------------------------------"
-sleep 3
-sudo apt-get install build-essential
-sudo apt-get install tcl8.5
-mkdir -p /tmp/redis
-wget http://download.redis.io/releases/redis-3.2.6.tar.gz -O /tmp/redis-3.2.6.tar.gz
-tar -xvzf /tmp/redis-3.2.6.tar.gz -C /tmp/redis --strip-components=1 && unlink /tmp/redis-3.2.6.tar.gz
-cd /tmp/redis && make
-make test
-sudo make install
-echo "-----------------------"
-echo "Installing the Redis Server. Just pressing Enter to accept the defaults for now."
-cd utils && sudo ./install_server.sh
-sudo update-rc.d redis_6379 defaults
-
+set _cwd = "$PWD"
 echo "--------------------------------------"
 echo "Installing the PostgreSQL dependencies..."
 echo "--------------------------------------"
 sleep 3
-# mkdir -p /tmp/psql
-# wget https://ftp.postgresql.org/pub/source/v9.5.5/postgresql-9.5.5.tar.gz -O /tmp/psql-9.5.5.tar.gz
-# tar -xvzf /tmp/psql-9.5.5.tar.gz -C /tmp/psql --strip-components=1 && unlink /tmp/redis-3.2.6.tar.gz
-# cd /tmp/psql && ./configure
-# make world
-# make check
-# sudo make install-world
-
-sudo apt-get install postgresql-9.5
+sudo service postgresql start
+pver=$( psql --version | sed -n 's/.* \([0-9.]*\)\.[0-9]*/\1/p' - )
+sudo service postgresql stop
+[ "$pver" != "9.5" ] && sudo apt-get --purge remove postgresql-$pver postgresql-contrib-$pver
+if [ ! -f /etc/apt/sources.list.d/pgdg.list ] || [ "$(sudo cat /etc/apt/sources.list.d/pgdg.list | grep 'apt.postgresql.org')" != "" ]; then
+    sudo sh -c 'echo "deb http://apt.postgresql.org/pub/repos/apt/ `lsb_release -cs`-pgdg main" >> /etc/apt/sources.list.d/pgdg.list'
+    wget --quiet -O - https://www.postgresql.org/media/keys/ACCC4CF8.asc | sudo apt-key add -
+fi
+sudo apt-get update
+sudo apt-get install postgresql-9.5 postgresql-contrib-9.5
 sudo sed -i "s/\(local *all *postgres *\)\w*/\1trust/g" /etc/postgresql/9.5/main/pg_hba.conf
 sudo sed -i "$ a local   all             tauchan                                trust" /etc/postgresql/9.5/main/pg_hba.conf
 
@@ -63,9 +27,49 @@ psql -c "CREATE ROLE tauchan ENCRYPTED PASSWORD 'md58040ed7558a08902c52968aafa41
 # UNENCRYPTED PASSWORD 'tauchan'
 createdb tauchan
 
-echo "---------------------"
-echo "Installing node dependencies"
-echo "---------------------"
-cd "$_cwd" && npm install
+# echo "------------------------------------------"
+# echo "Installing the GraphicsMagick dependencies..."
+# echo "Dependency downloading can take 5-10 minutes to complete. Please be patient."
+# echo "------------------------------------------"
+# sleep 3
+# mkdir -p /tmp/gm
+# wget ftp://ftp.graphicsmagick.org/pub/GraphicsMagick/1.3/GraphicsMagick-1.3.25.tar.gz -O /tmp/gm-1.3.25.tar.gz
+# tar -xzf /tmp/gm-latest.tar.gz -C /tmp/gm --strip-components=1 && unlink /tmp/gm-1.3.25.tar.gz
+# cd /tmp/gm && ./configure
+# make
+# make check
+# sudo make install
+
+# echo "----------------------------------"
+# echo "Installing the Redis dependencies..."
+# echo "Dependency tests can take 5-10 minutes to complete."
+# echo "There will be a couple times where user interaction is needed. Please stand by."
+# echo "----------------------------------"
+# sleep 3
+# sudo apt-get install build-essential
+# sudo apt-get install tcl8.5
+# mkdir -p /tmp/redis
+# wget http://download.redis.io/releases/redis-3.2.6.tar.gz -O /tmp/redis-3.2.6.tar.gz
+# tar -xvzf /tmp/redis-3.2.6.tar.gz -C /tmp/redis --strip-components=1 && unlink /tmp/redis-3.2.6.tar.gz
+# cd /tmp/redis && make
+# make test
+# sudo make install
+# echo "-----------------------"
+# echo "Installing the Redis Server. Just pressing Enter to accept the defaults for now."
+# cd utils && sudo ./install_server.sh
+# sudo update-rc.d redis_6379 defaults
+
+# mkdir -p /tmp/psql
+# wget https://ftp.postgresql.org/pub/source/v9.5.5/postgresql-9.5.5.tar.gz -O /tmp/psql-9.5.5.tar.gz
+# tar -xvzf /tmp/psql-9.5.5.tar.gz -C /tmp/psql --strip-components=1 && unlink /tmp/redis-3.2.6.tar.gz
+# cd /tmp/psql && ./configure
+# make world
+# make check
+# sudo make install-world
+
+# echo "---------------------"
+# echo "Installing node dependencies"
+# echo "---------------------"
+# cd "$_cwd" && npm install
 
 echo "Installation prerequisites setup has completed."
